+++ product/src/main/java/com/evo/product/domain/command/CreateOrUpdateProductCmd.java
@@ -17,12 +17,12 @@ public class CreateOrUpdateProductCmd {
     private UUID categoryId;
     private Map<String, String> description;
     private String introduce; // Introduce is a short description of the product(html)
-    private Double weight;
-    private Double length;
-    private Double width;
+    private int weight;
+    private int length;
+    private int width;
     private Long totalSold;
     private BigDecimal averageRating;
-    private Double height;
+    private int height;
     private Boolean hidden;
     List<CreateOrUpdateProductVariantCmd> productVariants;
     List<CreateOrUpdateProductImageCmd> productImages;
